Beginning of MVC refactoring
