+++ RFGraphMain.py
@@ -1,12 +1,13 @@
 #-*- coding: utf-8
 
 from PyQt4 import QtGui
-import ui
+import Optimisation
+import RFGraph_View as vw
 import sys
 
 
 qApp = QtGui.QApplication(sys.argv)
-aw = ui.ApplicationWindow()
+aw = vw.ApplicationWindow()
 aw.setWindowTitle('RFGraph')
 aw.show()
 sys.exit(qApp.exec_())
