+++ RFGraph_View.py
@@ -1,80 +1,9 @@
-#-*- coding: utf-8
-from math import *
-
-import numpy as np
 from PyQt4 import QtGui, QtCore
-from PyQt4.QtGui import *
-
+import numpy as np
 import data as fdata
 import graphs
-
-
-
-
-class MyTable(QTableWidget):
-    def __init__(self, data, *args):
-        QTableWidget.__init__(self, *args)
-        self.data = data
-
-        self.setmydata()
-        self.resizeColumnsToContents()
-        self.resizeRowsToContents()
-
-    def setmydata(self):
-        self.clear()
-        for n  in range(len(self.data)):
-            for m in range(len(self.data[n])):
-                newitem = QTableWidgetItem(self.data[n][m])
-                self.setItem(n,m,newitem)
-        self.setHorizontalHeaderLabels(['Complexité','Fitness','Equation'])
-
-class Optimisation(QDialog):
-    def __init__(self,parent=None):
-        super(Optimisation,self).__init__(parent)
-        god = QVBoxLayout(self)
-        layout = QHBoxLayout()
-        god.addLayout(layout)
-        s1 = QVBoxLayout()
-        s2 = QVBoxLayout()
-
-        popsize = QLineEdit()
-        concurr = QLineEdit()
-
-        s1.addWidget(QLabel("Pop size:"))
-        s1.addWidget(popsize)
-
-        s2.addWidget(QLabel("Concurrency:"))
-        s2.addWidget(concurr)
-
-
-        layout.addLayout(s1)
-        layout.addLayout(s2)
-
-        buttons = QDialogButtonBox(QDialogButtonBox.Ok | QDialogButtonBox.Cancel, QtCore.Qt.Horizontal, self)
-        buttons.accepted.connect(self.accept)
-        buttons.rejected.connect(self.reject)
-
-        god.addWidget(buttons)
-
-    def params(self):
-        return 0
-
-
-    @staticmethod
-    def get_params(Parent=None):
-        dialog = Optimisation()
-        result = dialog.exec_()
-        params = dialog.params()
-        return (result, params)
-
-#class Parameters:
-#    def __init__(self):
-#        self.mode_global = False
-#        self.lastNodeClicked = ''
-#        self.mode_cntrt = False
-#        self.click1 = ''
-#        self.click2 = ''
-
+from MyTable import MyTable
+from Optimisation import Optimisation
 
 class ApplicationWindow(QtGui.QMainWindow):
     def __init__(self):
@@ -213,7 +142,6 @@ class ApplicationWindow(QtGui.QMainWindow):
 
 
         if (not self.mode_cntrt):
-            #Ceci est un test pour le merge
             print('action:', nodeclicked)
             self.fitg.last_clicked = nodeclicked
             data_tmp = fdata.equacolOs[np.ix_(fdata.equacolOs[:, 2] == [nodeclicked], [0, 1, 3])]
@@ -253,4 +181,3 @@ class ApplicationWindow(QtGui.QMainWindow):
 
     def closeEvent(self, ce):
         self.fileQuit()
-
